Ensure code is 64-bit safe when accessing playerstat and entityState offsets

+++ code/qcommon/msg.c
@@ -627,7 +627,7 @@ typedef struct {
 } netField_t;
 
 // using the stringizing operator to save typing...
-#define NETF(x) #x, (int)&((entityState_t*)0)->x
+#define NETF(x) #x, (size_t) & ((entityState_t*)0)->x
 
 netField_t entityStateFields[] = {
     { NETF(pos.trTime), 32 },
@@ -937,7 +937,7 @@ void MSG_ReadDeltaEntity(msg_t* msg, entityState_t* from, entityState_t* to,
 }
 
 // using the stringizing operator to save typing...
-#define PSF(x) #x, (int)&((playerState_t*)0)->x
+#define PSF(x) #x, (size_t) & ((playerState_t*)0)->x
 
 netField_t playerStateFields[] = {
     { PSF(commandTime), 32 },
